refactor(digestive): T03 extract digestive data layer and fix layer violation
- Move Digestive entity to features/digestive/models/
- Move DigestiveWithSubmission domain model out of the UI page
  (tab_digestive_list.dart) into features/digestive/models/, removing the
  data -> UI dependency (problem P2)
- Move DigestiveRepository to features/digestive/repositories/
- Move digestive repository DI + query streams to
  features/digestive/repositories/digestive_providers.dart
- Move the submission-join query (undoneDigestivesWithSubmission) into a
  digestive use_case
- Update all references and regenerate build_runner outputs
