+++ lib/features/digestive/models/digestive_with_submission.dart
@@ -1,0 +1,17 @@
+import "../../submission/models/submission.dart";
+import "isar_digestive.dart";
+
+/// Digestive に紐づく Submission 情報を結合したドメインモデル。
+class DigestiveWithSubmission extends Digestive {
+  final Submission? submission;
+  final Digestive digestive;
+
+  DigestiveWithSubmission.fromObject(this.digestive, this.submission) {
+    id = digestive.id!;
+    submissionId = digestive.submissionId;
+    startAt = digestive.startAt;
+    minute = digestive.minute;
+    content = digestive.content;
+    done = digestive.done;
+  }
+}
