+++ lib/features/digestive/models/isar_digestive.dart
