+++ lib/features/submission/use_cases/delete_submission_use_case.dart
@@ -3,9 +3,9 @@ import "dart:developer";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../../../infrastructure/firebase_providers.dart";
-import "../../../providers/digestive_providers.dart";
-import "../../../repositories/digestive_repository.dart";
 import "../../../utils/types.dart";
+import "../../digestive/repositories/digestive_providers.dart";
+import "../../digestive/repositories/digestive_repository.dart";
 import "../../google_tasks/services/tasks_service.dart";
 import "../repositories/submission_providers.dart";
 import "../repositories/submission_repository.dart";
