+++ lib/pages/submission_detail_page.dart
@@ -8,12 +8,12 @@ import "package:url_launcher/url_launcher_string.dart";
 import "../components/digestive_detail_card.dart";
 import "../components/digestive_edit_bottom_sheet.dart";
 import "../components/submissions/formatted_date_remaining.dart";
+import "../features/digestive/models/isar_digestive.dart";
+import "../features/digestive/repositories/digestive_providers.dart";
 import "../features/submission/models/submission.dart";
 import "../features/submission/repositories/submission_providers.dart";
 import "../features/submission/repositories/submission_queries.dart";
 import "../infrastructure/firebase_providers.dart";
-import "../isar_db/isar_digestive.dart";
-import "../providers/digestive_providers.dart";
 import "../utils/ad_unit_ids.dart";
 import "../utils/ui.dart";
 import "../utils/utils.dart";
