+++ lib/sample_data.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 
+import "features/digestive/models/isar_digestive.dart";
 import "features/submission/models/submission.dart";
-import "isar_db/isar_digestive.dart";
 import "isar_db/isar_timetable.dart";
 import "utils/date_time_utils.dart";
