+++ test/features/submission/use_cases/delete_submission_use_case_test.dart
@@ -1,12 +1,12 @@
 import "package:flutter_test/flutter_test.dart";
 import "package:googleapis/tasks/v1.dart";
 import "package:mocktail/mocktail.dart";
+import "package:submon/features/digestive/models/isar_digestive.dart";
+import "package:submon/features/digestive/repositories/digestive_repository.dart";
 import "package:submon/features/google_tasks/services/tasks_service.dart";
 import "package:submon/features/submission/models/submission.dart";
 import "package:submon/features/submission/repositories/submission_repository.dart";
 import "package:submon/features/submission/use_cases/delete_submission_use_case.dart";
-import "package:submon/isar_db/isar_digestive.dart";
-import "package:submon/repositories/digestive_repository.dart";
 
 class MockSubmissionRepository extends Mock implements SubmissionRepository {}
